+++ src/react-native/javascript.ts
@@ -32,6 +32,7 @@ export async function addSentryInit({
     clack.log.warn(
       `Could not find main App file. Place the following code snippet close to the Apps Root component.`,
     );
+    Sentry.captureException('Could not find main App file.');
     await showCopyPasteInstructions(
       'App.js or _layout.tsx',
       getSentryInitColoredCodeSnippet(dsn, enableSessionReplay),
@@ -70,11 +71,15 @@ export async function addSentryInit({
       enableSessionReplay,
     });
 
-    clack.log.success(
-      `Added ${chalk.cyan('Sentry.init')} to ${chalk.cyan(jsRelativePath)}.`,
-    );
-
-    fs.writeFileSync(jsPath, newContent, 'utf-8');
+    try {
+      fs.writeFileSync(jsPath, newContent, 'utf-8');
+      clack.log.success(
+        `Added ${chalk.cyan('Sentry.init')} to ${chalk.cyan(jsRelativePath)}.`,
+      );
+    } catch (error) {
+      clack.log.error(`Error while writing ${jsPath}`);
+      Sentry.captureException(error);
+    }
   });
 
   Sentry.setTag('app-js-file-status', 'added-sentry-init');
@@ -193,11 +198,16 @@ export async function wrapRootComponent() {
   }
 
   traceStep('add-sentry-wrap', () => {
-    clack.log.success(
-      `Added ${chalk.cyan('Sentry.wrap')} to ${chalk.cyan(jsRelativePath)}.`,
-    );
-
-    fs.writeFileSync(jsPath, generateCode(mod.$ast).code, 'utf-8');
+    try {
+      fs.writeFileSync(jsPath, generateCode(mod.$ast).code, 'utf-8');
+      clack.log.success(
+        `Added ${chalk.cyan('Sentry.wrap')} to ${chalk.cyan(jsRelativePath)}.`,
+      );
+    } catch (error) {
+      clack.log.error(`Error while writing ${jsPath}`);
+      Sentry.captureException(error);
+      return;
+    }
   });
 
   Sentry.setTag('app-js-file-status', 'added-sentry-wrap');